Refactor move death scripts from event to death folder
